# internal/store/pgstore/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package pgstore

import (
	"net/netip"
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type AccessLog struct {
	ID            uuid.UUID  `json:"id"`
	InviteID      uuid.UUID  `json:"invite_id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	EnteredAt     time.Time  `json:"entered_at"`
	AuthorizedBy  *uuid.UUID `json:"authorized_by"`
}

type AccessRequest struct {
	ID            uuid.UUID  `json:"id"`
	UserID        uuid.UUID  `json:"user_id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	ApartmentID   uuid.UUID  `json:"apartment_id"`
	Status        string     `json:"status"`
	ReviewedBy    *uuid.UUID `json:"reviewed_by"`
	ReviewedAt    *time.Time `json:"reviewed_at"`
	CreatedAt     time.Time  `json:"created_at"`
	UpdatedAt     *time.Time `json:"updated_at"`
	Type          string     `json:"type"`
}

type Account struct {
	ID                    uuid.UUID  `json:"id"`
	UserID                uuid.UUID  `json:"user_id"`
	ProviderAccountID     string     `json:"provider_account_id"`
	ProviderID            string     `json:"provider_id"`
	PasswordHash          []byte     `json:"password_hash"`
	AccessToken           *string    `json:"access_token"`
	RefreshToken          *string    `json:"refresh_token"`
	AccessTokenExpiresAt  *int64     `json:"access_token_expires_at"`
	RefreshTokenExpiresAt *int64     `json:"refresh_token_expires_at"`
	Scope                 *string    `json:"scope"`
	IDToken               *string    `json:"id_token"`
	CreatedAt             time.Time  `json:"created_at"`
	UpdatedAt             *time.Time `json:"updated_at"`
}

type Announcement struct {
	ID            uuid.UUID  `json:"id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	AuthorID      *uuid.UUID `json:"author_id"`
	Title         string     `json:"title"`
	Content       string     `json:"content"`
	CreatedAt     time.Time  `json:"created_at"`
	UpdatedAt     *time.Time `json:"updated_at"`
}

type Apartment struct {
	ID            uuid.UUID  `json:"id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	Block         *string    `json:"block"`
	Number        string     `json:"number"`
	CreatedAt     time.Time  `json:"created_at"`
	UpdatedAt     *time.Time `json:"updated_at"`
}

type Bill struct {
	ID            uuid.UUID   `json:"id"`
	CondominiumID uuid.UUID   `json:"condominium_id"`
	ApartmentID   uuid.UUID   `json:"apartment_id"`
	BillType      string      `json:"bill_type"`
	ValueInCents  int64       `json:"value_in_cents"`
	DueDate       pgtype.Date `json:"due_date"`
	PaidAt        *time.Time  `json:"paid_at"`
	CreatedAt     time.Time   `json:"created_at"`
	UpdatedAt     *time.Time  `json:"updated_at"`
}

type Condominium struct {
	ID        uuid.UUID  `json:"id"`
	Name      string     `json:"name"`
	Cnpj      string     `json:"cnpj"`
	Address   string     `json:"address"`
	PlanType  string     `json:"plan_type"`
	CreatedAt time.Time  `json:"created_at"`
	UpdatedAt *time.Time `json:"updated_at"`
}

type CondominiumMember struct {
	ID            uuid.UUID  `json:"id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	UserID        uuid.UUID  `json:"user_id"`
	Role          string     `json:"role"`
	CreatedAt     time.Time  `json:"created_at"`
	UpdatedAt     *time.Time `json:"updated_at"`
}

type Invite struct {
	ID            uuid.UUID  `json:"id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	ApartmentID   uuid.UUID  `json:"apartment_id"`
	IssuedBy      uuid.UUID  `json:"issued_by"`
	GuestName     string     `json:"guest_name"`
	GuestType     string     `json:"guest_type"`
	Token         uuid.UUID  `json:"token"`
	StartsAt      time.Time  `json:"starts_at"`
	EndsAt        time.Time  `json:"ends_at"`
	RevokedAt     *time.Time `json:"revoked_at"`
	CreatedAt     time.Time  `json:"created_at"`
}

type Package struct {
	ID            uuid.UUID  `json:"id"`
	CondominiumID uuid.UUID  `json:"condominium_id"`
	ApartmentID   uuid.UUID  `json:"apartment_id"`
	ReceivedBy    uuid.UUID  `json:"received_by"`
	ReceivedAt    time.Time  `json:"received_at"`
	RecipientName *string    `json:"recipient_name"`
	PhotoUrl      *string    `json:"photo_url"`
	Status        string     `json:"status"`
	WithdrawnAt   *time.Time `json:"withdrawn_at"`
	WithdrawnBy   *uuid.UUID `json:"withdrawn_by"`
}

type Resident struct {
	ID            uuid.UUID `json:"id"`
	UserID        uuid.UUID `json:"user_id"`
	ApartmentID   uuid.UUID `json:"apartment_id"`
	Type          string    `json:"type"`
	IsResponsible bool      `json:"is_responsible"`
	CreatedAt     time.Time `json:"created_at"`
}

type Session struct {
	ID        uuid.UUID   `json:"id"`
	UserID    uuid.UUID   `json:"user_id"`
	Token     string      `json:"token"`
	ExpiresAt time.Time   `json:"expires_at"`
	IpAddress *netip.Addr `json:"ip_address"`
	UserAgent *string     `json:"user_agent"`
	CreatedAt time.Time   `json:"created_at"`
	UpdatedAt *time.Time  `json:"updated_at"`
}

type User struct {
	ID            uuid.UUID  `json:"id"`
	Name          string     `json:"name"`
	AvatarUrl     *string    `json:"avatar_url"`
	Email         string     `json:"email"`
	EmailVerified *time.Time `json:"email_verified"`
	CreatedAt     time.Time  `json:"created_at"`
	UpdatedAt     *time.Time `json:"updated_at"`
}

type UserDevice struct {
	ID         uuid.UUID `json:"id"`
	UserID     uuid.UUID `json:"user_id"`
	FcmToken   string    `json:"fcm_token"`
	Platform   *string   `json:"platform"`
	LastUsedAt time.Time `json:"last_used_at"`
	CreatedAt  time.Time `json:"created_at"`
}

type Verification struct {
	ID         uuid.UUID  `json:"id"`
	Identifier string     `json:"identifier"`
	Value      string     `json:"value"`
	ExpiresAt  time.Time  `json:"expires_at"`
	CreatedAt  time.Time  `json:"created_at"`
	UpdatedAt  *time.Time `json:"updated_at"`
}